getTruncateEmail util
