+++ apps/passport-client/new-components/shared/utils.ts
@@ -34,20 +34,3 @@ export const useExport = (): (() => Promise<void>) => {
     link.remove();
   }, [user, pcds, subscriptions]);
 };
-
-export const getTruncateEmail = (email: string, maxLength: number): string => {
-  const atIndex = email.lastIndexOf("@");
-  if (atIndex === -1 || email.length <= maxLength) {
-    return email;
-  }
-
-  const domain = email.slice(atIndex);
-  const localPart = email.slice(0, atIndex);
-  const availableLength = maxLength - domain.length - 3;
-
-  if (availableLength < 1) {
-    return email;
-  }
-
-  return `${localPart.slice(0, availableLength)}...${domain}`;
-};
